+++ source/src/sgii/sgiiBundle/Controller/TblEstadoProyectoController.php
@@ -28,6 +28,10 @@ class TblEstadoProyectoController extends Controller
      */
     public function indexAction()
     {
+        $security = $this->get('security');
+        if(!$security->autentication()){ return $this->redirect($this->generateUrl('login'));}
+        if(!$security->autorization($this->getRequest()->get('_route'))){ throw $this->createNotFoundException("Acceso denegado");}
+        
         $em = $this->getDoctrine()->getManager();
         $entities = $em->getRepository('sgiiBundle:TblEstadoProyecto')->findAll();
         return array( 'entities' => $entities );
@@ -46,6 +50,10 @@ class TblEstadoProyectoController extends Controller
      */
     public function showAction($id)
     {
+        $security = $this->get('security');
+        if(!$security->autentication()){ return $this->redirect($this->generateUrl('login'));}
+        if(!$security->autorization($this->getRequest()->get('_route'))){ throw $this->createNotFoundException("Acceso denegado");}
+        
         $em = $this->getDoctrine()->getManager();
         $entity = $em->getRepository('sgiiBundle:TblEstadoProyecto')->find($id);
 
@@ -74,6 +82,8 @@ class TblEstadoProyectoController extends Controller
     public function newAction(Request $request)
     {
         $security = $this->get('security');
+        if(!$security->autentication()){ return $this->redirect($this->generateUrl('login'));}
+        if(!$security->autorization($this->getRequest()->get('_route'))){ throw $this->createNotFoundException("Acceso denegado");}
         
         $entity = new TblEstadoProyecto();
         $form  = $this->createForm(new TblEstadoProyectoType(), $entity);
@@ -113,6 +123,8 @@ class TblEstadoProyectoController extends Controller
     public function editAction(Request $request, $id)
     {
         $security = $this->get('security');
+        if(!$security->autentication()){ return $this->redirect($this->generateUrl('login'));}
+        if(!$security->autorization($this->getRequest()->get('_route'))){ throw $this->createNotFoundException("Acceso denegado");}
         
         $em = $this->getDoctrine()->getManager();
         $entity = $em->getRepository('sgiiBundle:TblEstadoProyecto')->find($id);
@@ -159,6 +171,8 @@ class TblEstadoProyectoController extends Controller
     public function deleteAction(Request $request, $id)
     {
         $security = $this->get('security');
+        if(!$security->autentication()){ return $this->redirect($this->generateUrl('login'));}
+        if(!$security->autorization($this->getRequest()->get('_route'))){ throw $this->createNotFoundException("Acceso denegado");}
         
         $form = $this->createDeleteForm($id);
         $form->handleRequest($request);
@@ -170,12 +184,17 @@ class TblEstadoProyectoController extends Controller
             if (!$entity) {
                 throw $this->createNotFoundException('Unable to find TblEstadoProyecto entity.');
             }
-
-            $em->remove($entity);
-            $em->flush();
             
-            $security->setAuditoria('Eliminar estado de proyecto: '.$id. " - ".$entity->getEprEstadoProyecto());
-            $this->get('session')->getFlashBag()->add('alerts', array("type" => "information", "text" => "El estado de proyecto ha sido eliminado correctamente"));
+            $cantidad = $this->get('queries')->getCountEstadosProyecto($id);
+            if ($cantidad == 0) {
+                $em->remove($entity);
+                $em->flush();
+
+                $security->setAuditoria('Eliminar estado de proyecto: '.$id. " - ".$entity->getEprEstadoProyecto());
+                $this->get('session')->getFlashBag()->add('alerts', array("type" => "information", "text" => "El estado de proyecto ha sido eliminado correctamente"));
+            } else {
+                $this->get('session')->getFlashBag()->add('alerts', array("type" => "error", "text" => "El estado de proyecto no se puede eliminar porque hay proyectos con este."));
+            }
         }
         return $this->redirect($this->generateUrl('estadoproyecto'));
     }
