+++ source/src/sgii/sgiiBundle/Controller/TblTipoInvestigacionController.php
@@ -28,6 +28,10 @@ class TblTipoInvestigacionController extends Controller
      */
     public function indexAction()
     {
+        $security = $this->get('security');
+        if(!$security->autentication()){ return $this->redirect($this->generateUrl('login'));}
+        if(!$security->autorization($this->getRequest()->get('_route'))){ throw $this->createNotFoundException("Acceso denegado");}
+        
         $em = $this->getDoctrine()->getManager();
         $entities = $em->getRepository('sgiiBundle:TblTipoInvestigacion')->findAll();
         return array( 'entities' => $entities );
@@ -46,6 +50,10 @@ class TblTipoInvestigacionController extends Controller
      */
     public function showAction($id)
     {
+        $security = $this->get('security');
+        if(!$security->autentication()){ return $this->redirect($this->generateUrl('login'));}
+        if(!$security->autorization($this->getRequest()->get('_route'))){ throw $this->createNotFoundException("Acceso denegado");}
+        
         $em = $this->getDoctrine()->getManager();
         $entity = $em->getRepository('sgiiBundle:TblTipoInvestigacion')->find($id);
 
@@ -74,7 +82,9 @@ class TblTipoInvestigacionController extends Controller
     public function newAction(Request $request)
     {
         $security = $this->get('security');
-        
+        if(!$security->autentication()){ return $this->redirect($this->generateUrl('login'));}
+        if(!$security->autorization($this->getRequest()->get('_route'))){ throw $this->createNotFoundException("Acceso denegado");}
+
         $entity = new TblTipoInvestigacion();
         $form  = $this->createForm(new TblTipoInvestigacionType(), $entity);
         
@@ -113,6 +123,8 @@ class TblTipoInvestigacionController extends Controller
     public function editAction(Request $request, $id)
     {
         $security = $this->get('security');
+        if(!$security->autentication()){ return $this->redirect($this->generateUrl('login'));}
+        if(!$security->autorization($this->getRequest()->get('_route'))){ throw $this->createNotFoundException("Acceso denegado");}
         
         $em = $this->getDoctrine()->getManager();
         $entity = $em->getRepository('sgiiBundle:TblTipoInvestigacion')->find($id);
@@ -159,7 +171,9 @@ class TblTipoInvestigacionController extends Controller
     public function deleteAction(Request $request, $id)
     {
         $security = $this->get('security');
-        
+        if(!$security->autentication()){ return $this->redirect($this->generateUrl('login'));}
+        if(!$security->autorization($this->getRequest()->get('_route'))){ throw $this->createNotFoundException("Acceso denegado");}
+
         $form = $this->createDeleteForm($id);
         $form->handleRequest($request);
 
@@ -170,12 +184,17 @@ class TblTipoInvestigacionController extends Controller
             if (!$entity) {
                 throw $this->createNotFoundException('Unable to find TblTipoInvestigacion entity.');
             }
-
-            $em->remove($entity);
-            $em->flush();
             
-            $security->setAuditoria('Eliminar tipo de investigación: '.$id. " - ".$entity->getTinNombreTipo());
-            $this->get('session')->getFlashBag()->add('alerts', array("type" => "information", "text" => "El tipo de investigacion ha sido eliminado correctamente"));
+            $cantidad = $this->get('queries')->getCountTipoInvestigacion($id);
+            if ($cantidad == 0) {
+                $em->remove($entity);
+                $em->flush();
+
+                $security->setAuditoria('Eliminar tipo de investigación: '.$id. " - ".$entity->getTinNombreTipo());
+                $this->get('session')->getFlashBag()->add('alerts', array("type" => "information", "text" => "El tipo de investigacion ha sido eliminado correctamente"));
+            } else {
+                $this->get('session')->getFlashBag()->add('alerts', array("type" => "error", "text" => "El tipo de investigación no se puede eliminar porque hay proyectos con este."));
+            }
         }
         return $this->redirect($this->generateUrl('tipoinvestigacion'));
     }
